Support header_version 1 and 2
This adds support for the "recovery dtbo" and "dtb" parts in the header.

Whenever we add a feature from the later versions we update the header
version, but we never go backwards with versions. New headers start as
version 1, so will only become 1 or 2 as needed.

We also display some new information from v0 headers that were not shown
before (os version, header version), as well as the new data in the v1
and v2 headers.